Reorganize loop components. Add some information when setting a Loop URL.
