+++ src/app/components/Loop/LoopSetup.less
@@ -1,0 +1,11 @@
+.LoopSetup {
+  &-guide {
+    margin-bottom: 1rem;
+  }
+
+  &-form {
+    .ant-form-item {
+      margin-bottom: 0.5rem;
+    }
+  }
+}
